@@ -7,56 +7,62 @@
     "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
     "start": "nx serve",
     "build": "nx build",
-    "test": "nx test"
+    "test": "nx test",
+    "start-labguide": "concurrently \"nx serve labguide-server\" \"nx build labguide --watch\""
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.1.1",
-    "@angular/cdk": "13.1.1",
-    "@angular/common": "13.1.1",
-    "@angular/compiler": "13.1.1",
-    "@angular/core": "13.1.1",
-    "@angular/forms": "13.1.1",
-    "@angular/platform-browser": "13.1.1",
-    "@angular/platform-browser-dynamic": "13.1.0",
-    "@angular/router": "13.1.1",
-    "@nrwl/angular": "13.4.3",
+    "@angular/animations": "13.3.2",
+    "@angular/cdk": "13.3.2",
+    "@angular/common": "13.3.2",
+    "@angular/compiler": "13.3.2",
+    "@angular/core": "13.3.2",
+    "@angular/forms": "13.3.2",
+    "@angular/platform-browser": "13.3.2",
+    "@angular/platform-browser-dynamic": "13.3.2",
+    "@angular/router": "13.3.2",
+    "@nrwl/angular": "13.10.1",
     "bootstrap": "4.6.1",
-    "markdown-it": "^12.3.2",
-    "monaco-editor": "^0.31.1",
-    "ng-zorro-antd": "13.0.0",
-    "rxjs": "7.5.1",
+    "express": "4.17.3",
+    "markdown-it": "12.3.2",
+    "mongodb": "4.5.0",
+    "monaco-editor": "0.33.0",
+    "ng-zorro-antd": "13.1.1",
+    "rxjs": "7.5.5",
     "tslib": "2.3.1",
-    "zone.js": "0.11.4"
+    "zone.js": "0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.1",
-    "@angular-eslint/eslint-plugin": "~13.0.1",
-    "@angular-eslint/eslint-plugin-template": "~13.0.1",
-    "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.1.0",
-    "@angular/compiler-cli": "~13.1.0",
-    "@angular/language-service": "~13.1.0",
-    "@nrwl/cli": "13.4.1",
-    "@nrwl/cypress": "13.4.1",
-    "@nrwl/eslint-plugin-nx": "13.4.1",
-    "@nrwl/jest": "13.4.1",
-    "@nrwl/linter": "13.4.1",
-    "@nrwl/tao": "13.4.1",
-    "@nrwl/workspace": "13.4.1",
-    "@types/jest": "27.0.2",
-    "@types/markdown-it": "^12.2.3",
-    "@types/node": "^14.18.4",
-    "@typescript-eslint/eslint-plugin": "~5.3.0",
-    "@typescript-eslint/parser": "~5.3.0",
-    "cypress": "^9.1.0",
-    "eslint": "8.2.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
-    "prettier": "^2.3.1",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.4.3"
+    "@angular-devkit/build-angular": "13.3.2",
+    "@angular-eslint/eslint-plugin": "13.2.0",
+    "@angular-eslint/eslint-plugin-template": "13.2.0",
+    "@angular-eslint/template-parser": "13.2.0",
+    "@angular/cli": "13.3.2",
+    "@angular/compiler-cli": "13.3.2",
+    "@angular/language-service": "13.3.2",
+    "@nrwl/cli": "13.10.1",
+    "@nrwl/cypress": "13.10.1",
+    "@nrwl/eslint-plugin-nx": "13.10.1",
+    "@nrwl/express": "13.10.1",
+    "@nrwl/jest": "13.10.1",
+    "@nrwl/linter": "13.10.1",
+    "@nrwl/node": "13.10.1",
+    "@nrwl/tao": "13.10.1",
+    "@nrwl/workspace": "13.10.1",
+    "@types/express": "4.17.13",
+    "@types/jest": "27.4.1",
+    "@types/markdown-it": "12.2.3",
+    "@types/node": "17.0.23",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "5.18.0",
+    "cypress": "9.5.3",
+    "eslint": "8.13.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-cypress": "2.12.1",
+    "jest": "27.5.1",
+    "jest-preset-angular": "11.1.1",
+    "prettier": "2.6.2",
+    "ts-jest": "27.1.4",
+    "typescript": "4.6.3"
   }
 }
